Splits goal and lineup importers to separate files
