+++ trapp/command_line.py
@@ -3,11 +3,9 @@ from __future__ import absolute_import
 import argparse
 from trapp.database import Database
 from trapp.log import Log
-from trapp.importer import (
-    ImporterGoals,
-    ImporterLineups
-)
 from trapp.import_game import ImporterGames
+from trapp.import_goal import ImporterGoals
+from trapp.import_lineup import ImporterLineups
 from trapp.import_player import ImporterPlayers
